First commit message for this project. This is a mean-field solver for 2D translationally invariant Hamiltonian. Three modules and a Makefile have been included at this point.
So far, three modules were added, largely taken from a previous project (straintronics). These are alloc_dealloc (allocation and deallocation routines
for multidimensional arrays), init_routines (routines for initializing the values of arrays), and math_routines (various math functions).
Modifications were made to these, mostly just to make names and notation consistent. Unit testing was done, and seems to work as expected; however, I
wrote these a while ago, so confidence in the tests is not what it once was. Compiling and liking of each module and testing suite was automated in
the Makefile.

In the future, I would like to put some of these in a library; this would simplify maintenance of the code.