A (near-empty) driver file was added, and the Makefile was updated in consequence. Routines for complex floats were added to alloc_dealloc.
A driver was added; for now, it only serves as a placeholder. Its compilation and linking were automated in the Makefile. Alloc/dealloc routines for
complex floats were added to the alloc_dealloc module; these and the ones for complex doubles were unit-tested in the testing suite and seem to work
correctly.