A bit of code added to the driver. A first sample Hamiltonian module (ham1) is under development. Still a lot to do.
Arrays for the momentum grids and evals are allocated and initialized in the driver. Routines for evaluating the Hamiltonian at a given momentum and
evaluating the updated order parameter were added to the Hamiltonian module.

I'm still in the process of figuring out how to organize the program. Specifically, I haven't decided how to set up the parameter scan. The specific
parameters depend on the Hamiltonian -- one Hamiltonian's parameters may be different than anothers. I need to choose how to communicate this to the
driver. More broadly, I need to figure out a neat way of communicating all the necessary parameters from the Hamiltonian module to the driver.