Routine FermiEnerg(), which finds the Fermi energy given energy eigenvalues and number of occupied states, was added to math_routines.
The new routine uses the std::sort to order the energies in ascending order, and returns the energy at position filled_states. To do so, the array of
energies is copied -- this may be problematic for performance, especially regarding memory in the case of large arrays. The function prints out error
messages if filled_states is out of bounds. The routine was unit-tested in the module's testing suite, and seems to work as intended.