A module diag_routines was added. A LAPACKE-based routine for diagonalizing matrices in full storage layout was written.
The routine has not yet been tested. This is the simplest LAPACK routine for Hermitian matrices in full storage layout, but generally Relatively
Robust Representation is better, except for larger matrices in which case Divide and Conquer may be better (see documentation on Intel website). Since
the matrices we will diagonalize here will remain relatively small (8*8, or at worst 16*16), I don't think packed storage layout is necessary.

I am following a template similar to the diagonalization routines in the straintronics project, in the hopes of combining those with these into a
single library someday.