Diagonalization routine tested and integrated into the Makefile. struct added to the ham1 module to pass fixed parameters to the driver.
Compilation and unit testing for diag_routines module automated in the Makefile. The evals and evecs for a 4*4 Hermitian matrix were found and agree
with the results from numpy.

A struct called params_t is declared in ham1.h; its purpose is to transfer fixed parameters to the driver file. Accordingly, a function was added to
the ham1 module to assign parameter values to an instance of the struct. The driver was modified in consequence, that is, a instance of params_t is
declared and the function is called to assign its values, and these are used throughout the driver. Compilation and linking of the module and its
testing suite were automated in the Makefile, though the module has not yet been tested.

In the Makefile, INC_FLAGS was taken out of CXXFLAGS; this way, including locations in the path search will be done only when necessary.