In ham1 module, params_t was turned into a class; the constructor is now used to assign all the (const) member values.
The class params_t is intended to group parameters that any Hamiltonian will have. A constructor was programmed for the class, which assigns values to
the (constant) members (the values are set in ham1.cc, where the constructor is implemented). The driver was modified accordingly. Possible future
direction: make this class a singleton, since no more than one instance will ever be necessary.