Big change in design paradigm. Module added that defines a class for the momentum space and the associated energy bands.
I have realized that it makes most sense to define the Hamiltonian and the parameter space in the driver, at the cost of having a driver for every
Hamiltonian. Hence, we will delegate the creation of momentum space and the array for energy bands to another designated module. This module defines a
class which holds these arrays, as well as the necessary (constant) parameters. The module has been tested, and seems to behave as expected. Compiling
and building has been automated in the Makefile. Note that the object file must be linked to the alloc_dealloc.o and init_routines.o object files in
linking.