Much of the reorganization for the change in design paradigm was done. Work done on the (still not operational) driver_ham1.cc.
The file driver.cc was deleted; most of its contents now exists in the module kspace (from a few commits ago), which defines a class of momentum grids
and associated energy bands. The header ham1.h was deleted and the source code ham1.cc was renamed to driver_ham1.cc, reflecting its new role as a
driver file. It is still in this file that all the Hamiltonian parameters and routines are defined. The idea is that parameters that any Hamiltonian
file will define (such as those which specify momentum space, etc.) go into e.g. the kspace class, which forces the programmer to define them. The
driver is where the parameter space to be scanned is defined -- this must be the case, since the space is not the same for different Hamiltonians --
and where the work is done. I'm still developping it, so it is possible that other components get delegated to other modules in the form of routines
or classes.

The Makefile was reorganized in consequence. Now, each driver will have its own target (name of the executable to build) and its own internal clean
rule (all of which are called by the 'clean' target).