Modification to the evecs output in the diagonalization routine simple_zheev() in module diag_routines.
The routine simple_zheev() was modified so that evecs are written to a matrix z instead of (actually, in addition to) the matrix 'a' to be
diagonalized. This choice allows better consistency with other routines (which themselves output to a different matrix z) and other storage layouts
(for which the array 'a' cannot necessarily hold the eigenvectors).