The driver driver_ham1.cc has been brough to a functioning state. A lot of testing yet to be done. Output is printed, but not yet saved.
For each value of U defined in an array, the mean-field iteration is performed: at every loop, the Hamiltonian is diagonalized (using the previous
value of M) and the chemical potential is calculated; then, the eigenvectors (which we obtain by diagonalizing again), the energies, and the chemical
potential are used to calculate the new order parameter value. When the new and old OPs are the same within a tolerance of 1.e-15, the looping
terminates. It is not clear whether the answers are correct. To check, we could solve this simple model by hand (taking the field in the z direction
this time) and compare against the answers.