New module for NetCDF storage functionalities. A class was defined for creating simple NetCDF datasets. It is now used in driver_ham1 for IO.
The class newDS_t defined in the module creates a simple NetCDF dataset with n dimensions, n coordinate variables and m (other) variables spanning all
of these dimensions. To make this possible, the class constructor takes as arguments arrays of dimension names, dimension lengths, variable names,
etc, and likewise for the suggestively named functions WriteCoordVars() and WriteVars().

The constructor takes care of the 'define' stage of the NetCDF dataset: the dataset is created (its name is the time of creation and the user can
supply a path in the constructor arguments), the dimensions are created, and the variables (coordinate variables and others) are created; finally,
definition mode is ended. The variables are written using the member functions mentioned above, and the destructor releases the memory allocated by
the constructor and closes the dataset.

The module was unit-tested in its testing suite (though somewhat summarily), and compiling and linking were automated in the Makefile. The
driver_ham1.cc file was modified to use this class to output results in a NetCDF dataset.