A new driver (driver_ham2.cc) was written, largely modelled afer the first. It looks for Neel order in the Haldane-Hubbard model.
Here is a summary of major differences with respect to the first driver file: the unit cell in momentum space is carefully chosen to be the same as
the first Brillouin zone for a hexagonal lattice; the parameters entering the Hamiltonian are different; the parameter space we would like to search
has two dimensions U and t2 (though for now we are still treating t2 as fixed); the non-interacting and MFT Hamiltonians are obviously different (and
num_bands is 4); the routine to calculate the new magnetization is not quite the same. The data is saved as a NetCDF dataset, with the fixed
parameters indicated in the global attributes. Compilation and liking for this driver were automated in the Makefile under the target driver_ham2.

Next steps: set up the parameter study for t2 as well (possibly using a class to define the parameter space and variables on that space), and set up
plotting routines in Python.