In driver_ham2.cc, a class was defined to declare, allocate, and initialize arrays for the parameter space and order parameters. New bugs detected.
When an instance is declared, the arrays are declared, allocated, and initialized. They are the coordinate variables in the parameter space (only U
for now, though we will add t2) and the other variables, which are the order parameters (their converged values are to be saved here).

Newly noticed runtime bug: at the very end of the execution (after all the destructors have been called), we get "Abort trap: 6". This may be related
to reading arrays past the last value. Upon checking again, it seems this wasn't happening with driver_ham1.