Data saving was delegated to a method of the pspace_t class.
The new method which does this is SaveData(); it takes a string of metadata and a path as parameters. This makes sense because the parameters needed
to define the NetCDF dataset are precisely those characterizing the parameter space. The scope of the newDS_t object is limited to this method of the
pspace_t class.