Many warning in verbose warnings mode were taken care of.
A linebreak was added to the end of every document. A conflict between 'a' as an index and a global constant was resolved in driver_ham2.cc and
driver_ham1.cc. Type casting was made explicit in driver_ham2.cc, driver_ham1.cc, and init_routines.cc. In math_routines.cc, double literals were
changed to floats where appropriate. Unnecessary declaration of DateTime() was removed from nc_IO.h.