The driver driver_ham2.cc was modified to include t2 in the parameter space.
Similar to the case of the U axis, the class pspace_t defines a coordinate variable for t2. Accordingly, the space spanned by the OP M is now
two-dimensional (first index is t2 and second is U). The SaveData() method was modified accordingly. The do-while loop now loops over both dimensions
of parameter space, with the U dimension index changing faster. The only important difference in the loop is the argument to the function
Evaluate_ham(), which is now the appropriate component of pspace.t2_grid instead of a fixed value of t2.