Python driver and modules added for data plotting.
The module nc_IO.py contains a routine for reading a NetCDF; specifically, the same kind of simple NetCDF file created in this project -- one with n
dimensions, each with a coordinate variable, and m other variables spanning all the dimensions. The module plot_routines.py contains routines for
creating a 2D colorplot (one uses imshow; the other, still not finalized, uses pcolormesh). The file data_anal.py is a driver that calls the other
modules to load the data and plot it.