Attempt at parallelizing the parameter study in driver_ham2.cc using OpenMP.
We try parallelizing the loops over the parameter space, since these likely present a greater potential of efficiency. It's a happy coincidence that
organizing the code in classes and objects like I did makes things nicely tidy for parallelizing. The only (non-const) shared variable is the pspace
object; since every thread writes to a different part of it, this shouldn't be a problem. (Note: for some reason, I had to declare the kx_bounds and
ky_bounds arrays as shared even though they are const. Doesn't really matter, since we are not writing to them.) As I discovered, declaring arrays as
private doesn't do deep copies. Instead, they need to be declared and allocated (and later released) in the omp parallel code block. Hence, the kspace
object as well as the ham_array and evecs arrays are declared, allocated, and deallocated in the parallel code block. The Makefile was modified to
inlcude -fopenmp in the compilation and liking (and to use g++-7 instead of clang++).

Next step: use different stringstreams for each thread. It seems this may speed things up noticeably. Also, the output is not usefull as is.