Development of driver_ham3.cc was continued. The parameters and parameter spaces were redefined, and the function Dispersion() was updated.
For the parameters t1, t2, eps, and phi that are proper to each layer, a small class pars_t was created. This is useful because one axis of the
parameter space will be a scaling of the variables t1II, t2II, epsII, and tperp, and this scaling can be accomplished in the constructor. Other
parameters are defined as constants. Again, the parameter space is defined in the class pspace_t; its axes are the scaling factor alpha and the
Hubbard interaction strength U. The coordinate variable and the field variables (M and rhoI) are allocated in the constructor and deallocated in the
destructor. The method SaveData() saves these variables (along with attributes) to a NetCDF dataset. The function Dispersion (this misnomer ought to
be corrected in the future), which computes the 2*2 momentum-dependant Haldane Hamiltonian, was modified to use the pars_t class as an argument.