Development was continued on driver_ham3.cc. Work was done on the functions that assign Hamiltonian values and those that compute order parameters.
The function that computes the 2*2 k-deppendant Haldane Hamiltonian (renamed to Assign_h()) was made to take a pars_t as an argument. The function
that assigns the full 8*8 momentum-dependant Hamiltonian was updated -- the interface is ready but the implementation is not done. The functions
Compute_M_term() and Compute_rhoI_term() (the latter is new, the first was renamed) are tentatively operational, but there are almost certainly far
more efficient ways to implement them. Be careful about the denominator of the return value. Some preliminary changes were made to main() -- changing
some t2's to alphas, defining variables to hold rhoI, declaring pars_t instances, and outputting some rhoI values.