Development continued in driver_ham3.cc. Code compiles and executes, but for some parameters, calculation does not converge...
The lower triangle of the 8*8 momentum dependant matrix was entered according to pp. 23, 24, and 27 of my notes; the upper triangle should not matter
because it is not used by the diagonalization routine. Because it needs to be scaled, 'tperp' is defined in the loop from 'tperp_0' and 'alpha'.
Calculation and updating of the rhoI mean field was added into the do-while loop; said loop now terminates only if both mean fields have converged to
within tol. The program also prints rhoI values, etc. The global attributes have not yet been updated.

Even for those points for which the calculation converges, there seems to be strange behaviour in rhoI: the difference rhoIprime-rhoI does not
smoothly decrease, but instead sometimes jumps around by several orders of magnitude. Finally, for alpha=1, U=10, there seems to occur an endless
periodic loop. It is not clear to me what is going wrong, or even if the failure to converge is due to a programming bug.