In module math_routines, the function FermiEnerg() was rewritten without any copying of values and with better diagnostics output.
It was not necessary to keep the values of the input array untouched, since they are not reused in the calculation (more precisely, they are reused
but only after being recalculated). The optional diagnostics output now provides the minumum energy, the energy preceding the Fermi energy, the Fermi
energy itself, the one following it, and the max energy. The old function was kept and renamed FermiEnerg_cpy().

Note that in all cases, the FE is chosen to be the highest filled energy level; this choice agrees with the function nF0, which is one up to and
including zero energy.

Also, the function nF0() was modified so that no conversion betwen bool and float/double occurs (it's faster this way).