In the module diag_routines, an unsuccessful attempt was made at finding a better optimized diagonalization routine.
Despite promises in the documentation, the LAPACKE_zheevr() routine seems to not be faster than LAPACKE_zheev(), except maybe with evecs turned on and
for matrices of at least order 8. Still, it doesn't seem reliable enough to use, notably because there is an output (isuppz) whose meaning I can't
seem to figure out (seems to me like it should indicate 1 to MatOrder, but it doesn't).