Functionality added to driver_ham3.cc: the self-consistent search stops after a set number of loops. Attributes completed.
If the number of loops exceeds the set max without convergence, the mean field variable retains the array initialization value (chosen to be -99).

It seems that when the momentum grid is made more sparse (say by a factor of two), answers remain the same except near the transition, where certain
points don't converge (or converge very slowly). Maybe we can compensate by choosing a more sparse momentum grid and a higher max number of loops.