The number of failures to converge is now counted and reported in driver_ham3.cc. The max number of loops was increased to 3000.
In driver_ham3.cc, the variable that counts the number of failures, numfails, is a reduction variable because each threads adds to it. It is printed
at the end of the execution, and it is the return value of main().

In certain files, line returns were changed from std::endl to "\n", since the former will conserve output order accross threads when placed within a
string literal.