Loop-limiting functionality added to driver_ham2.cc (see previous commits).
If the number of loops exceeds 1000, the looping terminates and the mean field variable retains its initialization value. The number of failures to
converge is counted and reported.

Minor changes to driver_ham3.cc: definition of 'tol' (the tolerance) moved to outside of main(), with the other related parameters.

I just realized that holding all parameters of the Hamiltonian in a class would have allowed much more elegant code. Then, everything inside the for
loops which search the phase space could have been placed in a function. Before calling the functions, the phase space parameters could just be set to
the appropriate values. The class could even have a method for printing (or returning as a string) all the parameter values. This would even allow
further modularization: the functions for evaluating the Hamiltonian, computing the mean fields, etc. could be placed in one module, while the class
defining the phase space could be placed in another (or in the driver).