In driver_ham2.cc, the algorithm was updated to include the sublattice-symmetry-breaking mean field parameters.
There are now three mean-field parameters to be determined self-consistently: the sublattice-antisymmetric density rho_a, the sublattice-symmetric
magnetization mag_s, and the sublattice-antisymmetric magnetization mag_a (formerly M). This should have been the case from the start, since we are
including a term (eps) which exlicitely breaks sublattice symmetry. Since all the mean-field parameters are of the same order of magnitude, we use the
same tolerance for all of them. Use git diff for more details.

The saving of the additional mean-field parameters has not been implemented yet.