The saving of the new mean-field parameters was implemented. The number of loops required to converge is also saved.
The pspace class was modified to include arrays to hold the additional mean-field parameters as well as the number of loops that were required to
converge; these arrays are all initialized to the unlikely value -99. The MF parameter arrays are printed at the end of the execution, and all the
starting values are included in the dataset attributes.