Development of a separate module defining the Hamiltonian of driver_ham3.cc. This way, parameter study is separate from Hamiltonian, etc.
The design idea is to group everything not specific to a given parameter study is this new module. Specifically, a class is defined for the
Hamiltonian and the parameter values needed to compute it. The class contains an array to hold the Hamiltonian and a method for computing its value
for given momenta (allocated by the constructor and deallocated in the destructor). It also contains two pars_t objects, which hold parameters
relevant to the 2*2 Haldane Hamiltonian (This class was redesigned with a SetScaling() method, which reassigns parameter values given a scaling value;
to do this, the original values are held as private member variables.) The class ham3_t also has methods to calculate the MF expectation values.

A testing suite was created for the module; unit testing is still underway. All of the above was put into the Makefile.