Development with the new design concept was continued. Many parts of driver_ham3.cc were put into the ham3 module.
The Makefile was modified to link to the ham3.o object file when building driver_ham3.cc. All the constants and parameters relevant to the Hamiltonian
(or its momentum space) were put into the new module. A constant tperp_0 was also included so that tperp could be rescaled depending on parameters.
These parameters were made private and/or constant whenever possible (only 'a' could be made private actually). The function Assign_h() was made an
internal subroutine of ham3.cc; the functions Assign_ham(), ComputeTerm_mag(), and ComputeTerm_rhoI() were made methods of the ham3_t class. A method
was added for outputting metadata in the form of a string; this method is called within an 'omp single' construct to save the string in a shared
variable.

In the driver, the self-consistent looping was put into a function, IterativeSearch(). This way, it is conceptually separate from the parameter search
and can be called for single points in parameter space. This function takes a ham3_t instance, a kspace_t instance, and an evecs array as input; the
parameters of the ham3_t object must be adjusted beforehand. It also takes doubles corresponding to the MF parameters as input; these are used as the
starting values in the search and are as the output for the final values. The return is false if convergence was achieved and true if not. It would be
nice to make this routine a method of ham3_t, since it not directly associated to the parameter study. Also, we could then make ham3_t::mag and
ham3_t::rhoI private. The only issue is that we would have to include the kspace module in the ham3 header and source code.

The parameter study was put in its own routine, ParameterStudy(). At every step, the parameters in ham3 are adjusted as necessary and then
IterativeSearch() is called. Like before, the values are saved if convergence is reached; if not, a warning is given and the arrays retain their
initialization values. The routine returns the number of failures to converge.