The additional MF parameters were added to driver_ham3.cc and the ham3 module.
Changes were made to the files ham3.h, ham3.cc, and driver_ham3.cc. New methods for calculating the new MF parameters were added to the module ham3.
The chosen starting values were found to give good convergence for small U, but not for large U (equal to 8)... In addition to the new MF parameters,
the number of loops required to converge is also saved.