Further modularization: the evaluation of the 'out' MF parameters given the 'in' MF parameters is now done by a method of ham3_t.
To do this, ham_array, evecs, and kspace are made private members of ham3_t (allocated or constructed in the contructor). The MF parameter variables
used in Assign_ham() and other members are now also made private. In consequence, the modules kspace and diag_routines are now included in the
ham3.cc; the Makefile was updated in consequence. The module kspace is also included in ham3.h; this is not a problem because of preprocessor guards.