In the module ham3, the routine FixedPoint() was refined using a mixing scheme. Now, convergence seems far more robust.
Past a certain number of loops, the input value for the fuction evaluation is taken to be a linear mix between the previous output value and the
previous input value. As the number of loops crosses set thresholds, the proportion of the latter becomes 0.1, 0.2, 0.3, and 0.4. In particular, this
should resolve issues with the trajectory bouncing between values. In driver_ham3.cc, the testing parameters were reset to their previous values.