In driver_ham3.cc, the phase space in pspaceA_t was redefined in terms of lambda, the inverse of alpha. Wrong solution found at one point.
This way, only the energy parameters of parsI (the Haldane Hamiltonian parameters of the interacting layer) get scaled by lambda. Also, the
constructor of the pars_t class was removed, since it was unnecessary -- its role is fulfilled anyway by the SetScaling() method.

At one point in the test phase diagram, a ferromagnetic solution was found instead of an antiferromagnetic one.