New phase space study was written in driver_ham3.cc. This one studies L and U. Problems arriving at correct solution...
This new class pspaceB_t is similar to pspaceA_t, but with L and U as axes. Note that special care must be taken when setting lambda: we are defining
the coefficient in front of the Hubbard interaction as lambda*U, so here we choose to scale the interaction strength in the Hamiltonian. This way,
when plotting phase diagrams, the U axis shows the same range for different values of lambda. Of course, parsI is also scaled by lambda. In the
FixedPoint() routine, the thresholds for increasing chi were made smaller.

For large L, we sometimes find a ferromagnetic solution... Furthermore, the boundary between ferromagnetic and antiferromagnetic seems to depend on
starting values for MF parameters. I'm not sure what to do about this...