Module 'chempot' added. Module 'kspace' modified to suit the present needs.
The new module 'chempot' contains routines for computing the chemical potential. These were put in their own module because they are actually
implementations of root-finding algorithms. Right now, a bisection method algorithm and a Newton method algorithm have been implemented. The latter may
not be as stable as is required, so we will try stiking to the former. Note that the function evaluation in these solvers can be parallelized if
necessary.

The 'kspace' module was modified to be suitable for 3D space. To avoid implementing 4D arrays (which will actually have a large overhead if the last
dimension is short), we store the energies in a 1D array, making sure to track indexing. A function 'index()' is provided to do the indexing. Also, the
momentum grid is initialized to the Monkhorst-Pack scheme.