In the module kspace, the "evecs" member array was changed from a 1D array to a 3D array. Changes are still backward compatible.
The first index (globally) indexes the momenta, while the last two index the row and column of the "evecs" array. A routine is provided for getting
the momentum index. This change allows for the routines that calculate the MFs to be implemented more simply. The Makefile was updated to reflect the
use of the module "alloc" in "kspace.cc".