Several changes made to the data analysis scripts in the subdirectory data/.
A new routine for loading a whole dataset was added to the nc_IO.py module. Accordingly, the routine for loading a single variable was renamed.

A script data_anal_2.py was written that loads not only the requested variable, but also all variables defined on the same dimensions.

A script for plotting the difference in a var between two datasets was added.