The module ham4 was modified to assign the chemical potential to an attribute of the class ham4_t. Driver modified to save mu_.
Much like the free energy, the chemical potential is output by the method ComputeMFs() and is stored in the attribute mu_ in the method FixedPoint().
When the while loop of that method exits, the final value of the chemical potential is left in mu_ and can be saved from the driver.